+++ jc-web/src/main/java/com/realaicy/product/jc/modules/doccenter/repos/DocFileRepos.java
@@ -3,14 +3,13 @@ package com.realaicy.product.jc.modules.doccenter.repos;
 import com.realaicy.product.jc.modules.doccenter.model.DocFileRes;
 import com.realaicy.lib.core.orm.jpa.BaseRepository;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 2016/3/13.
  * XXX
  */
-public interface DocFileRepos extends BaseRepository<DocFileRes, Long> {
-
-
-
+public interface DocFileRepos extends BaseRepository<DocFileRes, BigInteger> {
 
 
 }
