+++ jc-web/src/main/java/com/realaicy/product/jc/modules/project/repos/PJApplyRepos.java
@@ -3,10 +3,12 @@ package com.realaicy.product.jc.modules.project.repos;
 import com.realaicy.lib.core.orm.jpa.BaseRepository;
 import com.realaicy.product.jc.modules.project.model.PJApply;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 2016/3/13.
  * XXX
  */
-public interface PJApplyRepos extends BaseRepository<PJApply, Long> {
+public interface PJApplyRepos extends BaseRepository<PJApply, BigInteger> {
 
 }
