+++ jc-web/src/main/java/com/realaicy/product/jc/modules/project/service/PJApplyService.java
@@ -3,10 +3,12 @@ package com.realaicy.product.jc.modules.project.service;
 import com.realaicy.lib.core.service.BaseService;
 import com.realaicy.product.jc.modules.project.model.PJApply;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 16/3/14.
  * xxx
  */
-public interface PJApplyService extends BaseService<PJApply, Long> {
+public interface PJApplyService extends BaseService<PJApply, BigInteger> {
 
 }
