+++ jc-web/src/main/java/com/realaicy/product/jc/modules/project/service/ProjectInfoService.java
@@ -3,10 +3,12 @@ package com.realaicy.product.jc.modules.project.service;
 import com.realaicy.lib.core.service.BaseService;
 import com.realaicy.product.jc.modules.project.model.ProjectInfo;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 16/3/14.
  * xxx
  */
-public interface ProjectInfoService extends BaseService<ProjectInfo, Long> {
+public interface ProjectInfoService extends BaseService<ProjectInfo, BigInteger> {
 
 }
