+++ jc-web/src/main/java/com/realaicy/product/jc/modules/project/service/impl/DefaultPJApplyService.java
@@ -5,12 +5,14 @@ import com.realaicy.product.jc.modules.project.model.PJApply;
 import com.realaicy.product.jc.modules.project.service.PJApplyService;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 16/3/14.
  * xxx
  */
 @Service
-public class DefaultPJApplyService extends DefaultBaseServiceImpl<PJApply, Long>
+public class DefaultPJApplyService extends DefaultBaseServiceImpl<PJApply, BigInteger>
         implements PJApplyService {
