+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/model/User.java
@@ -12,6 +12,7 @@ import com.realaicy.product.jc.modules.project.model.ProjectInfo;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 
@@ -20,20 +21,20 @@ import java.util.Set;
  */
 @Entity
 @Table(name = "jc_sys_user")
-public class User extends CommonDeletableEntity<Long> {
+public class User extends CommonDeletableEntity<BigInteger> {
 
 
     /**
      * 所属租户
      */
     @Column(name = "TENANTID")
-    private Long tenantID;
+    private BigInteger tenantID;
 
     /**
      * 所属组织
      */
     @Column(name = "ORGID")
-    private Long orgID;
+    private BigInteger orgID;
 
     /**
      * 用户名称
@@ -108,19 +109,19 @@ public class User extends CommonDeletableEntity<Long> {
         this.rolenames = rolenames;
     }
 
-    public Long getOrgID() {
+    public BigInteger getOrgID() {
         return orgID;
     }
 
-    public void setOrgID(Long orgID) {
+    public void setOrgID(BigInteger orgID) {
         this.orgID = orgID;
     }
 
-    public Long getTenantID() {
+    public BigInteger getTenantID() {
         return tenantID;
     }
 
-    public void setTenantID(Long tenantID) {
+    public void setTenantID(BigInteger tenantID) {
         this.tenantID = tenantID;
     }
