+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/repos/TenantRepos.java
@@ -3,11 +3,13 @@ package com.realaicy.product.jc.modules.system.repos;
 import com.realaicy.lib.core.orm.jpa.BaseRepository;
 import com.realaicy.product.jc.modules.system.model.Tenant;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 2016/3/13.
  * XXX
  */
-public interface TenantRepos extends BaseRepository<Tenant, Long> {
+public interface TenantRepos extends BaseRepository<Tenant, BigInteger> {
 
 
 }
