+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/repos/UserRepos.java
@@ -3,13 +3,14 @@ package com.realaicy.product.jc.modules.system.repos;
 import com.realaicy.lib.core.orm.jpa.BaseRepository;
 import com.realaicy.product.jc.modules.system.model.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
  * Created by realaicy on 2016/3/13.
  * XXX
  */
-public interface UserRepos extends BaseRepository<User, Long> {
+public interface UserRepos extends BaseRepository<User, BigInteger> {
 
     User findByUsername(String username);
