+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/service/UserService.java
@@ -5,13 +5,14 @@ import com.realaicy.product.jc.modules.system.model.User;
 import com.realaicy.product.jc.modules.system.model.vo.UserVO;
 import org.springframework.data.domain.PageRequest;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
  * Created by realaicy on 16/3/14.
  * xxx
  */
-public interface UserService extends BaseServiceWithVO<User, Long, UserVO> {
+public interface UserService extends BaseServiceWithVO<User, BigInteger, UserVO> {
 
     User findByName(String username);
