+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/service/impl/DefaultRoleService.java
@@ -18,7 +18,7 @@ import java.util.List;
  */
 @Service
 @Transactional
-public class DefaultRoleService extends DefaultBaseServiceImpl<Role, Long>
+public class DefaultRoleService extends DefaultBaseServiceImpl<Role, BigInteger>
         implements RoleService {
 
     @Override
