+++ jc-web/src/main/java/com/realaicy/product/jc/modules/system/service/impl/DefaultUserSecService.java
@@ -7,13 +7,15 @@ import com.realaicy.product.jc.modules.system.service.UserSecService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
+
 /**
  * Created by realaicy on 16/3/14.
  * xxx
  */
 @Service
 @Transactional
-public class DefaultUserSecService extends DefaultBaseServiceImpl<UserSec, Long>
+public class DefaultUserSecService extends DefaultBaseServiceImpl<UserSec, BigInteger>
         implements UserSecService {
 
     @Override
