+++ jc-web/src/main/java/com/realaicy/product/jc/realglobal/security/SpringSecurityPrincipal.java
@@ -2,6 +2,7 @@ package com.realaicy.product.jc.realglobal.security;
 
 import org.springframework.security.core.GrantedAuthority;
 
+import java.math.BigInteger;
 import java.util.Collection;
 import java.util.HashSet;
 
@@ -9,10 +10,10 @@ public final class SpringSecurityPrincipal extends RealUserDetails {
 
     private final String uuid;
 
-    public SpringSecurityPrincipal(final Long id, final String username, final String password, final String nickName, boolean enabled,
+    public SpringSecurityPrincipal(final BigInteger id, final String username, final String password, final String nickName, boolean enabled,
                                    boolean accountNonExpired, boolean credentialsNonExpired,
                                    boolean accountNonLocked,
-                                   final Collection<? extends GrantedAuthority> authorities, HashSet<String> realAuthorities, Long orgID, String orgCascadeID, final String uuidToSet) {
+                                   final Collection<? extends GrantedAuthority> authorities, HashSet<String> realAuthorities, BigInteger orgID, String orgCascadeID, final String uuidToSet) {
         super(id, username, password, nickName, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities, realAuthorities, orgID, orgCascadeID);
 
         uuid = uuidToSet;
